requirements.txt
django<1.11
psycopg2==2.6.2
django-extensions==1.7.1
python-openid==2.2.5
python-social-auth==0.2.20
requests-oauthlib==0.6.2
gunicorn==19.6.0
django-forms-bootstrap==3.0.1
rules==1.1.1
opbeat==3.3.4
pytest==2.9.2
pytest-django==2.9.1
tox==2.3.1
Pillow
django-stdimage==2.3.3
django-mailgun==0.9.1
opencage
pylama==7.0.9
icalendar==3.10
django-ical==1.4
ipython==5.1.0

# DRF
djangorestframework==3.4.6
markdown==2.6.6
django-filter==0.14.0
djangorestframework-jwt==1.8.0
django-rest-swagger==2.0.5
